+++ src/ll/orasql/scripts/oradrop.py
@@ -56,6 +56,10 @@ Options
 		If true, any exception that occurs while the database is read or written
 		will be ignored.
 
+	``--format`` : ``sql`` or ``pysql``
+		If ``--execute`` is not given, this determines the output format: Plain
+		SQL, or PySQL which can be piped into :mod:`ll.pysql`.
+
 	``--include`` : regexp
 		Only include objects in the output if their name contains the regular
 		expression.
@@ -90,6 +94,7 @@ def main(args=None):
 	p.add_argument("-x", "--execute", dest="execute", help="immediately execute the commands instead of printing them? (default %(default)s)", action=misc.FlagAction, default=False)
 	p.add_argument("-k", "--keepjunk", dest="keepjunk", help="Output objects with '$' in their name? (default %(default)s)", action=misc.FlagAction, default=False)
 	p.add_argument("-i", "--ignore", dest="ignore", help="Ignore errors? (default %(default)s)", default=False, action=misc.FlagAction)
+	p.add_argument(      "--format", dest="format", help="The output format (default %(default)s)", choices=("sql", "pysql"), default="sql")
 	p.add_argument(      "--include", dest="include", metavar="REGEXP", help="Include only objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
 	p.add_argument(      "--exclude", dest="exclude", metavar="REGEXP", help="Exclude objects whose name contains PATTERN (default: %(default)s)", type=re.compile)
 
@@ -152,7 +157,11 @@ def main(args=None):
 			if args.execute:
 				ddls.append((obj, ddl))
 			else:
-				stdout.write(ddl)
+				stdout.writeln(ddl.strip())
+				if args.format == "pysql":
+					stdout.writeln()
+					stdout.writeln("-- @@@")
+					stdout.writeln()
 
 	# Execute DDL
 	if args.execute: