+++ NEWS.rst
@@ -1,3 +1,11 @@
+Changes in 5.16 (released ??/??/2016)
+-------------------------------------
+
+* The scripts ``oracreate``, ``oradelete``, ``oradrop``, ``oragrant`` and
+	``orareindex`` have a new option :option:`--format`. The option value
+	``pysql`` switches the output format to PySQL output.
+
+
 Changes in 5.15.1 (released 03/21/2016)
 ---------------------------------------